Create scripty2.py
This will bin my smv files that are already of some size by a single number in x and y. It is parallel and fast. Still needs some fancy stuff like null outputs etc. Outputs are also tiff files without headers, but that is not a huge problem for now.